Refactoring of xAPI constants, extracted TrackerAsset subclasses
* official xAPI identifiers now in enums themselves
* mapping of id to names via TrackerUtils
* extracted TrackerEvent, TraceVerb
